bench/synth: fold the cell-rebound probe into the locals fixture and drop the two inlined-callee super fixtures
`locals_in_inlined_callee.py` gains a `probe_cell_rebound` arm reading one
cell through both the modelled mapping and LOAD_DEREF, and declares
`spec-folds=builtin_locals`; its standalone counterpart is removed.
`zero_arg_super_attr_inlined_callee.py` and
`zero_arg_name_bound_super_attr_inlined_callee.py` are removed with their
jitstats, and `zero_arg_name_bound_super_attr.py`'s comment now points at
the fixtures that remain.
