+++ pyre/bench/synth/locals_in_inlined_callee.wasm.jitstats
@@ -3,7 +3,7 @@ descr_set_absent=0
 descr_set_ambiguous=0
 descr_set_stale_absent=0
 fbw_blackhole_adopted_multi_frame=0
-fbw_blackhole_adopted_single_frame=0
+fbw_blackhole_adopted_single_frame=1
 fbw_escape_plain_fallback=0
 fbw_escape_plain_fallback_unclean=0
 fbw_midbody_latch_new_unjournaled=0
@@ -11,8 +11,8 @@ fbw_rolled_back_with_effects=0
 fbw_store_journal_rollback_failed=0
 field_pos_attached_misplaced=0
 field_pos_spec_misplaced=0
-guard_failures=3
+guard_failures=4
 internal_compile_panics=0
-loops_aborted=0
-loops_compiled=3
+loops_aborted=1
+loops_compiled=4
 retraces_compiled=0
