--- pyre/bench/synth/zero_arg_name_bound_super_attr_inlined_callee.wasm.jitstats
@@ -1,18 +1,0 @@
-bridges_compiled=0
-descr_set_absent=0
-descr_set_ambiguous=0
-descr_set_stale_absent=0
-fbw_blackhole_adopted_multi_frame=0
-fbw_blackhole_adopted_single_frame=0
-fbw_escape_plain_fallback=0
-fbw_escape_plain_fallback_unclean=0
-fbw_midbody_latch_new_unjournaled=0
-fbw_rolled_back_with_effects=0
-fbw_store_journal_rollback_failed=0
-field_pos_attached_misplaced=0
-field_pos_spec_misplaced=0
-guard_failures=1
-internal_compile_panics=0
-loops_aborted=0
-loops_compiled=1
-retraces_compiled=0
